move get_surface_backend tests to test_surf_plotting
